refactor: clarify dispatcher HTTP module
